refactor package name of s2counter_loader.
