+++ s2counter_loader/src/main/scala/org/apache/s2graph/counter/loader/models/DefaultCounterModel.scala
@@ -1,8 +1,6 @@
-package org.apache.s2graph.models
+package org.apache.s2graph.counter.loader.models
 
+import org.apache.s2graph.counter.models.CounterModel
 import org.apache.s2graph.spark.config.S2ConfigFactory
 
-/**
- * Created by hsleep(honeysleep@gmail.com) on 15. 6. 8..
- */
 case object DefaultCounterModel extends CounterModel(S2ConfigFactory.config)
