+++ s2counter_loader/src/main/scala/org/apache/s2graph/counter/loader/stream/ExactCounterStreaming.scala
@@ -1,30 +1,27 @@
-package org.apache.s2graph.counter.stream
+package org.apache.s2graph.counter.loader.stream
 
 import kafka.serializer.StringDecoder
-import org.apache.s2graph.counter.core.CounterFunctions
+import org.apache.s2graph.counter.config.S2CounterConfig
+import org.apache.s2graph.counter.loader.config.StreamingConfig
+import org.apache.s2graph.counter.loader.core.CounterFunctions
 import org.apache.s2graph.spark.config.S2ConfigFactory
 import org.apache.s2graph.spark.spark.{WithKafka, SparkApp, HashMapParam}
 import org.apache.spark.streaming.Durations._
 import org.apache.spark.streaming.kafka.KafkaRDDFunctions.rddToKafkaRDDFunctions
 import org.apache.spark.streaming.kafka.{HasOffsetRanges, StreamHelper}
-import s2.counter.core.CounterFunctions
-import spark.spark.WithKafka
-
 import scala.collection.mutable.{HashMap => MutableHashMap}
+import scala.language.postfixOps
 
-/**
- * Created by hsleep(honeysleep@gmail.com) on 15. 6. 19..
- */
-object RankingCounterStreaming extends SparkApp with WithKafka {
+object ExactCounterStreaming extends SparkApp with WithKafka {
   lazy val config = S2ConfigFactory.config
   lazy val s2Config = new S2CounterConfig(config)
   lazy val className = getClass.getName.stripSuffix("$")
 
   lazy val producer = getProducer[String, String](StreamingConfig.KAFKA_BROKERS)
 
-  val inputTopics = Set(StreamingConfig.KAFKA_TOPIC_COUNTER_TRX)
+  val inputTopics = Set(StreamingConfig.KAFKA_TOPIC_COUNTER)
   val strInputTopics = inputTopics.mkString(",")
-  val groupId = buildKafkaGroupId(strInputTopics, "ranking_v2")
+  val groupId = buildKafkaGroupId(strInputTopics, "counter_v2")
   val kafkaParam = Map(
 //    "auto.offset.reset" -> "smallest",
     "group.id" -> groupId,
@@ -51,23 +48,19 @@ object RankingCounterStreaming extends SparkApp with WithKafka {
     // make stream
     val stream = streamHelper.createStream[String, String, StringDecoder, StringDecoder](ssc, inputTopics)
     stream.foreachRDD { (rdd, ts) =>
-      // for at-least once semantic
-      val nextRdd = {
-        CounterFunctions.makeRankingRdd(rdd, sc.defaultParallelism).foreachPartition { part =>
-          // update ranking counter
-          CounterFunctions.updateRankingCounter(part, acc)
-        }
-        rdd
-      }
+      val offsets = rdd.asInstanceOf[HasOffsetRanges].offsetRanges
 
-      streamHelper.commitConsumerOffsets(nextRdd.asInstanceOf[HasOffsetRanges])
-//      CounterFunctions.makeRankingRdd(rdd, offsets.length).foreachPartitionWithIndex { (i, part) =>
-//        // update ranking counter
-//        CounterFunctions.updateRankingCounter(part, acc)
-//
-//        // commit offset range
-//        streamHelper.commitConsumerOffset(offsets(i))
-//      }
+      val exactRDD = CounterFunctions.makeExactRdd(rdd, offsets.length)
+
+      // for at-least once semantic
+      exactRDD.foreachPartitionWithIndex { (i, part) =>
+        // update exact counter
+        val trxLogs = CounterFunctions.updateExactCounter(part.toSeq, acc)
+        CounterFunctions.produceTrxLog(trxLogs)
+
+        // commit offset range
+        streamHelper.commitConsumerOffset(offsets(i))
+      }
     }
 
     ssc.start()
