--- s2counter_loader/src/main/scala/s2/models/DefaultCounterModel.scala
@@ -1,8 +1,0 @@
-package s2.models
-
-import org.apache.s2graph.spark.config.S2ConfigFactory
-
-/**
- * Created by hsleep(honeysleep@gmail.com) on 15. 6. 8..
- */
-case object DefaultCounterModel extends CounterModel(S2ConfigFactory.config)
